requirements.txt
torch>=2.0
torchvision
efficientnet_pytorch
pandas
numpy
scikit-learn
matplotlib
seaborn
tqdm
Pillow
jupyter